Changed Method of Checking for Ranked Rod
The old method of checking between fishing_rods was to check to see if
the Display Name == Ranked Rod(ignoreCase). This for some reason chucked
out errors. And could be exploited by naming a default fishing_rod
Ranked Rod with it having no lore. Now we just check to see if the rod
in question has a lore. If so run Ranked Rod protocol.